+++ src/ErrorN.ts
@@ -1,0 +1,7 @@
+"use strict";
+
+class ErrorN extends Error {
+    public code: string;
+}
+
+export default ErrorN;