+++ src/types/Bower.d.ts
@@ -1,0 +1,5 @@
+declare interface Bower {
+    logger: BowerLogger;
+    config: BowerConfig;
+    version: string;
+}