+++ src/types/BowerLogger.d.ts
@@ -1,0 +1,13 @@
+declare interface BowerLogger {
+    error: Function;
+    conflict: Function;
+    warn: Function;
+    action: Function;
+    info: Function;
+    debug: Function;
+    log: Function;
+    prompt: Function;
+    pipe: Function;
+    geminate: Function;
+    intercept: Function;
+}