+++ src/types/ProGetConf.d.ts
@@ -1,0 +1,8 @@
+declare interface ProGetConf {
+    apiKeyMapping: Array<ProGetApiConf>;
+}
+
+declare interface ProGetApiConf {
+    server: RegExp;
+    key: string;
+}