+++ src/types/ReleaseTags.d.ts
@@ -1,0 +1,4 @@
+declare interface ReleaseTags {
+    target: string;
+    version: string;
+}